// frontend/lib/api/generated/model/characterNetworkResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AI小说创作系统
 * 作者主导的长篇小说创作平台：常驻创作 Agent、固定生成工作流与可编排的复合任务
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterNetworkResponseNetworkAnalysis } from './characterNetworkResponseNetworkAnalysis';
import type { CharacterRelationshipResponse } from './characterRelationshipResponse';
import type { CharacterResponse } from './characterResponse';

/**
 * 角色关系网络响应
 */
export interface CharacterNetworkResponse {
  characters: CharacterResponse[];
  network_analysis?: CharacterNetworkResponseNetworkAnalysis;
  novel_id: number;
  relationships: CharacterRelationshipResponse[];
}